Refine vdev-skipping logic
Skip a vdev in the !data->mounted case when a pool's guid has already
been determined and the vdev's pool guid doesn't match.

This fixes pool detection in a number of cases in which multiple pools
exist.  For example, previously if a system had a 3-disk raidz followed
by a single-disk pool, scanning of the 4th disk would cause the first
3-disk raidz pool to not be visible.